requirements.txt
requests
apprise
python-dotenv
